🚧 [hotkeys] dynamic scope loading
